+++ abstract/ActivityBlock.js
@@ -60,6 +60,14 @@ export class ActivityBlock extends Block {
           this.$['*history'] = [];
         }
       });
+
+      if (this.has('*modalActive')) {
+        this.sub('*modalActive', (modalActive) => {
+          if (!modalActive && this.activityType === this.$['*currentActivity']) {
+            this.$['*currentActivity'] = null;
+          }
+        });
+      }
     }
   }